Add Python3.14 Support

+++ opflexagent/test/test_gbp_ovs_agent.py
@@ -109,6 +109,7 @@ class TestGBPOpflexAgent(base.OpflexTestBase):
             mock.patch('neutron.agent.common.ovs_lib.BaseOVS.get_bridges'),
             mock.patch('oslo_service.loopingcall.FixedIntervalLoopingCall',
                 new=MockFixedIntervalLoopingCall),
+            mock.patch('opflexagent.opflex_notify.worker', return_value=None),
             mock.patch('opflexagent.gbp_agent.GBPOpflexAgent.'
                 '_report_state')]
 
@@ -145,7 +146,8 @@ class TestGBPOpflexAgent(base.OpflexTestBase):
         agent.bridge_manager.trunk_rpc = mock.Mock()
         agent.of_rpc.get_gbp_details = mock.Mock()
         agent.port_manager.of_rpc.request_endpoint_details_list = mock.Mock()
-        agent.notify_worker.terminate()
+        if agent.notify_worker is not None:
+            agent.notify_worker.terminate()
 
     def test_port_unbound_snat_cleanup(self):
         self.agent.int_br = mock.Mock()
